refactor(config): move subagent config to objects folder
- Move `SubAgentConfiguration` to `Config/Objects`
- This improves project structure and organization.

Commit generated by The Saturn Pipeline
